+++ src/Install/Installer.php
@@ -25,6 +25,7 @@ use Mollie\Exception\CouldNotInstallModule;
 use Mollie\Factory\ModuleFactory;
 use Mollie\Handler\ErrorHandler\ErrorHandler;
 use Mollie\Utility\MultiLangUtility;
+use Mollie\Utility\TabTranslationUtility;
 use OrderState;
 use PrestaShopException;
 use Tab;
@@ -251,7 +252,7 @@ class Installer implements InstallerInterface
         $languages = Language::getLanguages(true);
 
         foreach ($languages as $language) {
-            $moduleTab->name[$language['id_lang']] = $this->module->l($name, false, $language['iso_code']);
+            $moduleTab->name[$language['id_lang']] = TabTranslationUtility::getTabName($this->module, $name, $language['iso_code']);
         }
 
         if (!$moduleTab->save()) {